+++ src/main/java/com/ams/backend/repository/AuditedRepository.java
@@ -3,6 +3,6 @@ package com.ams.backend.repository;
 import com.ams.backend.entity.Audited;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditedRepository extends JpaRepository<Audited, Long> {
+public interface AuditedRepository extends JpaRepository<Audited, Integer> {
 
 }
